# Conflicts:
#	CHANGELOG
#	SHE_PPT/python/SHE_PPT/she_frame.py
#	SHE_PPT/python/SHE_PPT/she_frame_stack.py
#	SHE_PPT/python/SHE_PPT/she_image.py
#	SHE_PPT/python/SHE_PPT/testing/utility.py
#	SHE_PPT/tests/python/she_image_test.py

+++ SHE_PPT/python/SHE_PPT/table_formats/she_galaxy_population_priors.py
@@ -63,12 +63,12 @@ class SheGalaxyPopulationPriorsFormat(SheTableFormat):
         # Column names and info
 
         self.ID = self.set_column_properties(
-            "SHE_GALPOP_OBJECT_ID", dtype = ">a20", fits_dtype = "PA(20)")
+            "SHE_GALPOP_OBJECT_ID", dtype=">a20", fits_dtype="PA(20)")
 
         self.ra = self.set_column_properties(
-            "SHE_GALPOP_RA", comment = "deg", dtype = ">f8", fits_dtype = "D")
+            "SHE_GALPOP_RA", comment="deg", dtype=">f8", fits_dtype="D")
         self.dec = self.set_column_properties(
-            "SHE_GALPOP_DEC", comment = "deg", dtype = ">f8", fits_dtype = "D")
+            "SHE_GALPOP_DEC", comment="deg", dtype=">f8", fits_dtype="D")
 
         self.zp = self.set_column_properties(
             "SHE_GALPOP_Z_PHOT")
@@ -76,158 +76,158 @@ class SheGalaxyPopulationPriorsFormat(SheTableFormat):
             "SHE_GALPOP_Z_SPEC")
 
         self.iab = self.set_column_properties(
-            "SHE_GALPOP_I_AB", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_I_AB", comment="mag",
+            dtype=">f4", fits_dtype="E")
 
         self.vab = self.set_column_properties(
-            "SHE_GALPOP_V_AB", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_V_AB", comment="mag",
+            dtype=">f4", fits_dtype="E")
 
         self.beta_0 = self.set_column_properties(
-            "SHE_GALPOP_BETA_0", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_BETA_0", comment="mag",
+            dtype=">f4", fits_dtype="E")
         self.e_bulge = self.set_column_properties(
-            "SHE_GALPOP_E_BULGE", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_E_BULGE", comment="mag",
+            dtype=">f4", fits_dtype="E")
         self.e_disk = self.set_column_properties(
-            "SHE_GALPOP_E_DISK", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_E_DISK", comment="mag",
+            dtype=">f4", fits_dtype="E")
         self.e_galaxy = self.set_column_properties(
-            "SHE_GALPOP_E_GALAXY", comment = "mag",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_E_GALAXY", comment="mag",
+            dtype=">f4", fits_dtype="E")
 
         self.sers_sing_fit = self.set_column_properties(
             "SHE_GALPOP_N_SERSIC_SINGLE_FIT",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.sers_two_fit = self.set_column_properties(
             "SHE_GALPOP_N_SERSIC_BULGE_TWO_COMP_FIT",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.disk_length = self.set_column_properties(
-            "SHE_GALPOP_H_DISK_ARCSEC", comment = "arcsec",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_H_DISK_ARCSEC", comment="arcsec",
+            dtype=">f4", fits_dtype="E")
         self.bulge_hlr = self.set_column_properties(
             "SHE_GALPOP_REFF_BULGE",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.tau_b_rst = self.set_column_properties(
             "SHE_GALPOP_TAU_B_REST",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.x = self.set_column_properties(
             "SHE_GALPOP_X_CENTRE_PIXEL",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.y = self.set_column_properties(
             "SHE_GALPOP_Y_CENTRE_PIXEL",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.tilt = self.set_column_properties(
-            "SHE_GALPOP_THETA_DEG", comment = "deg",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_THETA_DEG", comment="deg",
+            dtype=">f4", fits_dtype="E")
         self.rotation = self.set_column_properties(
-            "SHE_GALPOP_POS_ANGLE_DEG", comment = "deg",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_POS_ANGLE_DEG", comment="deg",
+            dtype=">f4", fits_dtype="E")
         self.one_component_chi2 = self.set_column_properties(
             "SHE_GALPOP_CHI2_SINGE_FIT",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.two_component_chi2 = self.set_column_properties(
             "SHE_GALPOP_CHI2_TWO_COMP_FIT",
-            dtype = ">f4", fits_dtype = "E")
+            dtype=">f4", fits_dtype="E")
         self.two_component_dof = self.set_column_properties(
             "SHE_GALPOP_NDEG_TWO_COMP_FIT",
-            dtype = ">i8", fits_dtype = "K")
+            dtype=">i8", fits_dtype="K")
         self.bulge_fd_f435w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F435W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F435W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f435w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F435W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F435W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f606w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F606W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F606W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f606w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F606W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F606W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f775w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F775W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F775W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f775w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F775W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F775W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f814w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F814W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F814W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f814w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F814W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F814W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f850lp = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F850LP", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F850LP", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f850lp_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F850LP", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F850LP", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f105w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F105W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F105W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f105w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F105W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F105W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f125w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F125W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F125W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f125w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F125W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F125W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f160w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_BULGE_NJY_F160W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_BULGE_NJY_F160W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.bulge_fd_f160w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F160W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_BULGE_NJY_F160W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f435w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F435W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F435W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f435w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F435W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F435W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f606w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F606W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F606W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f606w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F606W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F606W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f775w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F775W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F775W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f775w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F775W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F775W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f814w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F814W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F814W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f814w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F814W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F814W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f850lp = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F850LP", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F850LP", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f850lp_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F850LP", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F850LP", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f105w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F105W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F105W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f105w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F105W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F105W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f125w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F125W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F125W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f125w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F125W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F125W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f160w = self.set_column_properties(
-            "SHE_GALPOP_FLUX_DISK_NJY_F160W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_DISK_NJY_F160W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
         self.disk_fd_f160w_err = self.set_column_properties(
-            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F160W", comment = "nJy",
-            dtype = ">f4", fits_dtype = "E")
+            "SHE_GALPOP_FLUX_ERR_DISK_NJY_F160W", comment="nJy",
+            dtype=">f4", fits_dtype="E")
 
         self._finalize_init()
